Oracle migration step 1b (TriRef + MeshIDRelation): accessors -> native types
The TriRefs()/MeshIDTransforms() accessors on both Impl and MutableImpl
now return mesh.TriRef (int fields) and a native meshIDRelation (geom.Mat3x4
transform) instead of leaking bridge.TriRef/bridge.MeshIDRelation. The
int32->int field-width change rippled to every SetTriRefs-building caller
(int32() casts at the storage boundary) and every meshID-keyed map (int keys).

After this the accessor API leaks no bridge types; the only remaining
bridge.TriRef/bridge.MeshIDRelation in production are the two const-Impl
read-side converters in impl_storage.go, which step 2 removes.

+++ go/csg_tree.go
@@ -146,10 +146,10 @@ func compose(leaves []*Impl) *MutableImpl {
 		offset := int32(i) * meshIDCounterSnapshot
 		refs := node.TriRefs()
 		for t, r := range refs {
-			triMeshID[triIndices[i]+t] = r.MeshID + offset
-			triOriginalID[triIndices[i]+t] = r.OriginalID
-			triFaceID[triIndices[i]+t] = r.FaceID
-			triCoplanarID[triIndices[i]+t] = r.CoplanarID
+			triMeshID[triIndices[i]+t] = int32(r.MeshID) + offset
+			triOriginalID[triIndices[i]+t] = int32(r.OriginalID)
+			triFaceID[triIndices[i]+t] = int32(r.FaceID)
+			triCoplanarID[triIndices[i]+t] = int32(r.CoplanarID)
 		}
 	}
